refactor!: do not copy JS lib to platform project
