+++ .eslintignore
@@ -1,4 +1,3 @@
-bin/node_modules/*
 templates/project/*
 tests/spec/unit/fixtures/*
