+++ lib/BridgingHeader.js
