+++ lib/create.js
@@ -21,9 +21,9 @@ const path = require('path');
 const fs = require('fs-extra');
 const xmlescape = require('xml-escape');
 const { CordovaError, events } = require('cordova-common');
-const pkg = require('../../package');
+const pkg = require('../package');
 
-const ROOT = path.join(__dirname, '../..');
+const ROOT = path.join(__dirname, '..');
 
 /**
  * Creates a new iOS project with the following options:
@@ -101,11 +101,6 @@ class ProjectCreator {
         const srcScriptsDir = path.join(ROOT, 'bin/templates/scripts/cordova');
         const destScriptsDir = this.projectPath('cordova');
         fs.copySync(srcScriptsDir, destScriptsDir);
-
-        const nodeModulesDir = path.join(ROOT, 'node_modules');
-        if (fs.existsSync(nodeModulesDir)) {
-            fs.copySync(nodeModulesDir, path.join(destScriptsDir, 'node_modules'));
-        }
     }
 
     expandTokens () {
