+++ lib/list-devices
