+++ lib/list-emulator-images
