+++ lib/listDevices.js
