+++ lib/listEmulatorImages.js
