+++ lib/plugman/pluginHandlers.js
