+++ lib/prepare.js
