+++ tests/spec/unit/build.spec.js
@@ -21,7 +21,7 @@ const path = require('path');
 const fs = require('fs-extra');
 const rewire = require('rewire');
 const { CordovaError, events } = require('cordova-common');
-const build = rewire('../../../bin/templates/scripts/cordova/lib/build');
+const build = rewire('../../../lib/build');
 
 describe('build', () => {
     const testProjectPath = path.join('/test', 'project', 'path');
@@ -385,7 +385,7 @@ describe('build', () => {
 
         beforeEach(() => {
             // rewire causes some issues so for these tests, we will require instead.
-            buildRequire = require('../../../bin/templates/scripts/cordova/lib/build');
+            buildRequire = require('../../../lib/build');
             spyOn(events, 'emit');
         });
